some fixes

+++ scrumboard/static/js/app/views.js
@@ -149,6 +149,7 @@ var BoardSetView = Backbone.View.extend({
     },
     new_board : function () {
         var board_title = window.prompt("Title ?");
+        if (!board_title) return;
         var board = new Board({"title" : board_title})
         board.save()
         this.model.fetch()
@@ -169,26 +170,30 @@ var BoardSetView = Backbone.View.extend({
     }
 })
 
+
 var BoardView = Backbone.View.extend({
     el : $("#board-info"),
     initialize : function () {
-        this.model.bind("change", this.render, this)
+        this.model.bind("change", this.render, this);
     },
     events : {
         "click .destroy" : 'destroy',
-        "click .change" : 'change',
+        "click .change" : 'change'
     },
     change : function () {
         var title = window.prompt("Title ?", this.model.get("title"))
+        if (!title) return;
         this.model.set("title", title)
         this.model.save({ success: this.render })
     },
     destroy : function () {
-        this.model.destroy({
-            success : function () {
-                window.location = '/';
-            }
-        });
+        if (window.confirm('Are you sure ?')) {
+            this.model.destroy({
+                success : function () {
+                    window.location = '/';
+                }
+            });
+        }
 
     },
     render : function () {
